% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp.R
\name{cp}
\alias{cp}
\title{Copy to clipboard function}
\usage{
cp(x, rn = FALSE)
}
\arguments{
\item{x}{A dataframe, matrix or other data object}

\item{rn}{Rownames are not copied when argument is set to FALSE (default)}
}
\description{
Copies data object to clipboard
}
\details{
This function copies data object to clipboard for easy paste in Word, Excel
or email.
}
\examples{
copydata <- matrix(ncol=10,nrow=10,data=rnorm(100,100,10))
cp(copydata)

}
\author{
Yves R. Sagaert
}
