% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pt}
\alias{pt}
\title{pt function}
\usage{
pt(data)
}
\arguments{
\item{data}{A dataframe or matrix}
}
\value{
A data frame or matrix
}
\description{
Gives orderly and limited head of the dataframe or matrix.
}
\details{
This function gives a subsample of the data for visual inspection. The
number of columns is limited to 10, the number of rows is limited to 5.
}
\examples{
bigdataframe <- matrix(ncol=100,nrow=100,data=rnorm(10000,100,10))
pt(data = bigdataframe)

}
\author{
Yves R. Sagaert
}
