% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforecasting_package.R
\docType{package}
\name{salesforecasting}
\alias{salesforecasting}
\alias{salesforecasting-package}
\title{salesforecasting: A package for sales forecasting in a business setting.}
\description{
This package provides functionalities for sales forecasting in a business
setting.
}
\author{
Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
